small refactors

+++ BMS-LIB/Inc/Models/Battery/Battery.hpp
@@ -19,7 +19,7 @@ struct voltage_register_group {
 
 class Battery {
 public:
-	static const uint8_t CELLS = 6;
+	static const int CELLS = 6;
 	static const uint8_t MAX_SOC_DIFFERENCE = 100;
 	static const uint8_t MIN_CELL_VOLTAGE = 3.4;
 	static const uint8_t MAX_CELL_VOLTAGE = 4.2;
